pyproject.toml
[tool.poetry]
name = "spamnesty"
version = "0.1.0"
description = ""
authors = ["Stavros Korokithakis <hi@stavros.io>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
bleach = "*"
brotlipy = "*"
coverage = "*"
django = "==1.11.*,>=1.11.0"
django-annoying = "*"
django-bootstrap3 = "*"
django-extensions = "*"
django-nose = "*"
django-redis = "*"
faker = "*"
hypothesis = {extras = ["django"], version = "*"}
ipython = "*"
numpy = "*"
psycopg2 = "*"
raven = "*"
scipy = "*"
shortuuid = "*"
sklearn = "*"
spintax = "*"
uwsgi = "*"
werkzeug = "*"
whitenoise = "*"

[tool.poetry.dev-dependencies]
flake8 = "*"
flake8-docstrings = "*"
mypy-lang = "*"
pep8 = "*"
pep8-naming = "*"
pip-tools = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
